Jobs configuration moved to new subpackage ...eva.pipeline.jobs
