+++ eva-pipeline/src/main/java/embl/ebi/variation/eva/pipeline/jobs/VariantAggregatedConfiguration.java
@@ -1,4 +1,4 @@
-package embl.ebi.variation.eva;
+package embl.ebi.variation.eva.pipeline.jobs;
 
 
 import org.slf4j.Logger;
